+++ src/Calculators/Dominion/ImprovementCalculator.php
@@ -31,14 +31,26 @@ class ImprovementCalculator
      * @param string $improvementType
      * @return float
      */
-    public function getImprovementMultiplierBonus(Dominion $dominion, string $improvementType): float
+    public function getImprovementMultiplierBonus(Dominion $dominion, string $improvementType, bool $secondary = false): float
     {
         $improvementPoints = $dominion->{'improvement_' . $improvementType};
         $totalLand = $this->landCalculator->getTotalLand($dominion);
 
         $multiplier = $this->getImprovementMaximum($improvementType)
-            * (1 - exp(-$improvementPoints / ($this->getImprovementCoefficient($improvementType) * $totalLand + 15000)))
-            * $this->getImprovementMultiplier($dominion);
+            * (1 - exp(-$improvementPoints / ($this->getImprovementCoefficient($improvementType) * $totalLand + 15000)));
+
+        // Ignore Masonries
+        if (!($secondary && $improvementType == 'spires')) {
+            $multiplier *= $this->getImprovementMultiplier($dominion);
+        }
+
+        if ($secondary) {
+            if ($improvementType == 'spires') {
+                $multiplier *= 0.5;
+            } elseif ($improvementType == 'harbor') {
+                $multiplier *= 1.25;
+            }
+        }
 
         return round($multiplier, 4);
     }